cli: drop tasks add/ready/stale/orphans/preflight; hide tasks dispatch
- 'add' was a literal alias for 'create'
- ready/stale/orphans now flags on 'tasks list'
- preflight is the composition '--stale=7 --orphans'
- dispatch parent/worker hidden — they're hub-only verbs

Audit ref: docs/code-review/2026-04-28-hipp-audit.md §1
Ousterhout ref: docs/code-review/2026-04-28-ousterhout-plan-audit.md §Phase 1
