+++ docs/site/layouts/index.html
@@ -559,7 +559,7 @@ $ bones --help</pre>
     <pre id="quickstart">$ brew install danmestas/tap/bones
 $ cd my-project
 $ bones up                                    # workspace + scaffold + leaf + hub
-$ bones tasks add "first task" --files src/foo.go
+$ bones tasks create "first task" --files src/foo.go
 $ bones tasks list</pre>
   </div>
   <p>Walkthrough: <a href="/docs/quickstart/">/docs/quickstart</a> · Concepts: <a href="/docs/concepts/">/docs/concepts</a> · CLI reference: <a href="/docs/reference/cli/">/docs/reference/cli</a> · Skills reference: <a href="/docs/reference/skills/">/docs/reference/skills</a> · Architecture: <a href="/docs/architecture/">/docs/architecture</a> · Source: <a href="https://github.com/danmestas/bones">github.com/danmestas/bones</a>.</p>
